Parameters verification server side

+++ App/Controllers/GuideController.php
@@ -40,10 +40,10 @@ class GuideController extends AControllerBase
         $images = array();
 
         if ($id != null) {
-            $sections = Section::getAll("guide_id = $id");
+            $sections = Section::getAll("guide_id = ?", [$id]);
             foreach ($sections as $section) {
                 $section_id = $section->getID();
-                $images_with_id = Image::getAll("section_id = $section_id");
+                $images_with_id = Image::getAll("section_id = ?", [$section_id]);
                 $images += array_merge($images, $images_with_id);
             }
         }
@@ -57,8 +57,10 @@ class GuideController extends AControllerBase
 
     public function save(): Response
     {
-
         $id = $this->request()->getValue("id");
+        if (!is_numeric($id)) {
+            return $this->redirect($this->url("home.index", ["id" => $id]));
+        }
         $post = $this->request()->getPost();
 
         $keys = array_keys($post);
@@ -66,12 +68,14 @@ class GuideController extends AControllerBase
 
         foreach ($section_ids as $section_id) {
             $section_id = str_replace("s_id_", "", $section_id);
-            $section = Section::getOne($section_id);
-            $text = $post["text_$section_id"];
-            $header = $post["s_header_$section_id"];
-            $section->setText($text == null ? null : $text);
-            $section->setHeader($header);
-            $section->save();
+            if (is_numeric($section_id)) {
+                $section = Section::getOne($section_id);
+                $text = $post["text_$section_id"];
+                $header = $post["s_header_$section_id"];
+                $section->setText($text == null ? null : $text);
+                $section->setHeader($header);
+                $section->save();
+            }
         }
 
         $new_section_ids = preg_grep("/new_s_\d+$/", $keys);
@@ -79,37 +83,42 @@ class GuideController extends AControllerBase
 
         foreach ($new_section_ids as $section_id) {
             $section_id = str_replace("new_s_", "", $section_id);
-            $header = $post["new_header_$section_id"];
-            $text = $post["new_text_$section_id"];
-            $section = new Section();
-            $section->setGuideId($id);
-            $section->setText($text == null ? null : $text);
-            $section->setHeader($header);
-            $section->setName($header);
-            $section->setOrder($order);
-            $section->setParentSection(null);
-            $section->save();
-            $order++;
+            if (is_numeric($section_id)) {
+                $header = $post["new_header_$section_id"];
+                $text = $post["new_text_$section_id"];
+                $section = new Section();
+                $section->setGuideId($id);
+                $section->setText($text == null ? null : $text);
+                $section->setHeader($header);
+                $section->setName($header);
+                $section->setOrder($order);
+                $section->setParentSection(null);
+                $section->save();
+                $order++;
+            }
         }
 
-        return $this->redirect($this->url("guide.index", ["id" => $this->request()->getValue("id")]));
+        return $this->redirect($this->url("guide.index", ["id" => $id]));
     }
 
     public function delete(): Response
     {
         $id = $this->request()->getValue("id");
-        $sections = Section::getAll("guide_id = $id");
 
-        foreach ($sections as $section) {
-            $section_id = $section->getID();
-            $images = Image::getAll("section_id = $section_id");
-            foreach ($images as $image) {
-                $image->delete();
+        if (is_numeric($id)) {
+            $sections = Section::getAll("guide_id = ?", [$id]);
+
+            foreach ($sections as $section) {
+                $section_id = $section->getID();
+                $images = Image::getAll("section_id = $section_id");
+                foreach ($images as $image) {
+                    $image->delete();
+                }
+                $section->delete();
             }
-            $section->delete();
-        }
 
-        Guide::getOne($id)->delete();
+            Guide::getOne($id)->delete();
+        }
 
         return $this->redirect($this->url("home.index"));
     }
